billing_service/requirements.txt
fastapi==0.101.1
gunicorn==20.1.0
uvicorn==0.21.0
orjson==3.9.4
psycopg2==2.9.6
asyncpg==0.28.0
aiohttp==3.8.5
alembic==1.11.2
pydantic==2.1.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.20
confluent-kafka==2.2.0
yookassa==2.4.0